chore(ui): delete 'stores' directory, move client & profiles into utils (no longer acting as stores)
