+++ ui/src/components/SearchEverythingInput.vue
@@ -82,7 +82,7 @@
 import { isHashtag, TAG_SYMBOLS } from "@/utils/tags";
 import { SearchResult, SearchResultOption } from "@/types/types";
 import { ROUTES, PATH, router } from "@/router";
-import { useSearchProfiles } from "@/stores/profiles";
+import { useSearchProfiles } from "@/utils/profiles";
 import { showError } from "@/utils/notification";
 import { AppAgentClient, encodeHashToBase64 } from "@holochain/client";
 import { QSelect, QSelectOption, QIcon, QItem, QItemSection } from "quasar";
